+++ src/Modern.WindowKit/RawEventGrouper.cs
@@ -2,10 +2,8 @@
 using System;
 using System.Collections.Generic;
 //using Modern.WindowKit.Collections.Pooled;
-using Modern.WindowKit.Input;
 using Modern.WindowKit.Input.Raw;
 using Modern.WindowKit.Threading;
-//using JetBrains.Annotations;
 
 namespace Modern.WindowKit;
 
@@ -19,7 +17,7 @@ internal class RawEventGrouper : IDisposable
     private readonly Action<RawInputEventArgs> _eventCallback;
     private readonly Queue<RawInputEventArgs> _inputQueue = new();
     private readonly Action _dispatchFromQueue;
-    readonly Dictionary<long, RawTouchEventArgs> _lastTouchPoints = new();
+    readonly Dictionary<long, RawPointerEventArgs> _lastTouchPoints = new();
     RawInputEventArgs? _lastEvent;
 
     public RawEventGrouper(Action<RawInputEventArgs> eventCallback)
@@ -39,7 +37,7 @@ internal class RawEventGrouper : IDisposable
     private void DispatchFromQueue()
     {
         while (true)
-        {
+    {
             if(_inputQueue.Count == 0)
                 return;
 
@@ -49,7 +47,7 @@ internal class RawEventGrouper : IDisposable
                 _lastEvent = null;
             
             if (ev is RawTouchEventArgs { Type: RawPointerEventType.TouchUpdate } touchUpdate)
-                _lastTouchPoints.Remove(touchUpdate.TouchPointId);
+                _lastTouchPoints.Remove(touchUpdate.RawPointerId);
 
             _eventCallback?.Invoke(ev);
 
@@ -62,7 +60,7 @@ internal class RawEventGrouper : IDisposable
                 return;
             }
         }
-    }
+            }
     
     public void HandleEvent(RawInputEventArgs args)
     {
@@ -87,12 +85,12 @@ internal class RawEventGrouper : IDisposable
             && lastPointerEvent.Type is RawPointerEventType.Move or RawPointerEventType.TouchUpdate)
         {
             if (args is RawTouchEventArgs touchEvent)
-            {
-                if (_lastTouchPoints.TryGetValue(touchEvent.TouchPointId, out var lastTouchEvent))
+        {
+                if (_lastTouchPoints.TryGetValue(touchEvent.RawPointerId, out var lastTouchEvent))
                     MergeEvents(lastTouchEvent, touchEvent);
                 else
                 {
-                    _lastTouchPoints[touchEvent.TouchPointId] = touchEvent;
+                    _lastTouchPoints[touchEvent.RawPointerId] = touchEvent;
                     AddToQueue(touchEvent);
                 }
             }
@@ -105,10 +103,10 @@ internal class RawEventGrouper : IDisposable
         {
             _lastTouchPoints.Clear();
             if (args is RawTouchEventArgs { Type: RawPointerEventType.TouchUpdate } touchEvent)
-                _lastTouchPoints[touchEvent.TouchPointId] = touchEvent;
+                _lastTouchPoints[touchEvent.RawPointerId] = touchEvent;
         }
         AddToQueue(args);
-    }
+        }
 
     private static IReadOnlyList<RawPointerPoint> GetList() => new List<RawPointerPoint>();
     private static readonly Func<IReadOnlyList<RawPointerPoint>> s_getListDelegate = GetList;
