+++ src/Modern.WindowKit/RawTouchEventArgs.cs
@@ -1,15 +1,23 @@
-﻿namespace Modern.WindowKit.Input.Raw
-{
+﻿using System;
+
+namespace Modern.WindowKit.Input.Raw
+    {
     public class RawTouchEventArgs : RawPointerEventArgs
     {
         public RawTouchEventArgs(IInputDevice device, ulong timestamp, IInputRoot root,
             RawPointerEventType type, Point position, RawInputModifiers inputModifiers,
-            long touchPointId) 
+            long rawPointerId) 
             : base(device, timestamp, root, type, position, inputModifiers)
         {
-            TouchPointId = touchPointId;
+            RawPointerId = rawPointerId;
         }
 
-        public long TouchPointId { get; set; }
+        public RawTouchEventArgs(IInputDevice device, ulong timestamp, IInputRoot root,
+            RawPointerEventType type, RawPointerPoint point, RawInputModifiers inputModifiers,
+            long rawPointerId)
+            : base(device, timestamp, root, type, point, inputModifiers)
+        {
+            RawPointerId = rawPointerId;
+        }
     }
 }
